+++ extras/doc/CHANGELOG.md
@@ -1,7 +1,11 @@
 # CHANGELOG GC9D01
 
 * Version 1.0.1 02-2026
-	* first version 
+  * first version
 * Version 1.0.2 02-2026
-	* fix for 90 and 270 degree rotation for 40x160 single gate displays see gihub issue 1
- 
+  * fix for 90 and 270 degree rotation for 40x160 single gate displays see github issue 1
+* Version 1.1.0 05-2026
+  * Added Support for 50x160 single gate display
+  * Added Support for 60x160 single gate display
+  * Added two new parameters to TFTInitScreenSize() :portrait Column start offset
+  & portrait Row start offset required for new displays added.